Improvements of Interface and code generator + LLM reading doc
I've organize a bit the github and improved the code of llamag_interface.py. The LLM can now acces the doc and therefore give better responses when not included in the Q&A
